Update versioning and refactor game UI components
- Bump version to 1.0.74 and adjust assembly versions.
- Introduce `IGameConsoleUI` interface for better abstraction.
- Refactor `BasePlayerClient` to utilize the new interface.
- Implement `IGameConsoleUI` in `GameConsoleUI` class.
- Update method signatures to use `byte` for board positions.
- Enhance null safety with nullable types in `BasePlayerClient`.
- Add unit tests for `GameConsoleUI` and `BoardRenderer`.
- Update project files for .NET 8.0 compatibility.
- Revise solution structure to include `Player.Tests`.
- Improve error handling and logging throughout the codebase.

+++ src/TicTacToe.Tournament.BasePlayer/Interfaces/IBot.cs
@@ -4,9 +4,9 @@ namespace TicTacToe.Tournament.BasePlayer.Interfaces;
 
 public interface IBot
 {
-    Task<(int row, int col)> MakeMoveAsync(Guid matchId, Mark[][] board);
+    Task<(byte row, byte col)> MakeMoveAsync(Guid matchId, Mark[][] board);
     void OnMatchStarted(Guid matchId, Guid playerId, Guid opponentId, Mark mark, bool starts);
-    void OnOpponentMoved(Guid matchId, int row, int col);
+    void OnOpponentMoved(Guid matchId, byte row, byte col);
     void OnMatchEnded(GameResult result);
     void OnBoardUpdated(Guid matchId, Mark[][] board);
 }
