+++ src/TicTacToe.Tournament.BasePlayer/Interfaces/IPlayerStrategy.cs
@@ -4,6 +4,6 @@ namespace TicTacToe.Tournament.BasePlayer.Interfaces
 {
     public interface IPlayerStrategy
     {
-        (int row, int col) MakeMove(Mark[][] board);
+        (byte row, byte col) MakeMove(Mark[][] board);
     }
 }
